+++ resources/js/bootstrap.js
@@ -25,10 +25,10 @@ window.axios.defaults.headers.common['X-Requested-With'] = 'XMLHttpRequest';
 
 // window.Echo = new Echo({
 //     broadcaster: 'pusher',
-//     key: import.meta.env.VITE_PUSHER_APP_KEY,
-//     wsHost: import.meta.env.VITE_PUSHER_HOST ?? `ws-${import.meta.env.VITE_PUSHER_APP_CLUSTER}.pusher.com`,
-//     wsPort: import.meta.env.VITE_PUSHER_PORT ?? 80,
-//     wssPort: import.meta.env.VITE_PUSHER_PORT ?? 443,
-//     forceTLS: (import.meta.env.VITE_PUSHER_SCHEME ?? 'https') === 'https',
+//     key: process.env.VITE_PUSHER_APP_KEY,
+//     wsHost: process.env.VITE_PUSHER_HOST ?? `ws-${process.env.VITE_PUSHER_APP_CLUSTER}.pusher.com`,
+//     wsPort: process.env.VITE_PUSHER_PORT ?? 80,
+//     wssPort: process.env.VITE_PUSHER_PORT ?? 443,
+//     forceTLS: (process.env.VITE_PUSHER_SCHEME ?? 'https') === 'https',
 //     enabledTransports: ['ws', 'wss'],
 // });