+++ resources/views/app.blade.php
@@ -1,0 +1,16 @@
+<!DOCTYPE html>
+<html class="h-full">
+
+<head>
+  <meta charset="utf-8" />
+  <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0" />
+  <link href="{{ mix('/css/app.css') }}" rel="stylesheet" />
+  <script src="{{ mix('/js/app.js') }}" defer></script>
+  @inertiaHead
+</head>
+
+<body class="h-full">
+  @inertia
+</body>
+
+</html>