+++ tests/Feature/ExampleTest.php
@@ -1,21 +1,7 @@
 <?php
 
-namespace Tests\Feature;
+it('asserts true is true', function () {
+    $this->assertTrue(true);
 
-// use Illuminate\Foundation\Testing\RefreshDatabase;
-use Tests\TestCase;
-
-class ExampleTest extends TestCase
-{
-    /**
-     * A basic test example.
-     *
-     * @return void
-     */
-    public function test_the_application_returns_a_successful_response()
-    {
-        $response = $this->get('/');
-
-        $response->assertStatus(200);
-    }
-}
+    expect(true)->toBeTrue();
+});